- Cart screen states (loading, error, empty). - Cart detail. - Decimal usage (to avoid losing precision).

+++ lib/extensions/double_extensions.dart
@@ -1,3 +1,5 @@
-extension DoubleExtensions on double {
+import 'package:decimal/decimal.dart';
+
+extension DoubleExtensions on Decimal {
   String get formatWithCurrency => '£${toStringAsFixed(2)}';
 }
