+++ lib/modules/product/data/model/product_model.dart
@@ -1,4 +1,5 @@
 import 'package:checkout/modules/product/domain/entity/product.dart';
+import 'package:decimal/decimal.dart';
 import 'package:meta/meta.dart';
 
 @immutable
@@ -6,7 +7,7 @@ final class ProductModel {
   final String sku;
   final String? imageUrl;
   final String name;
-  final double price;
+  final Decimal price;
 
   const ProductModel({
     required this.sku,
